+++ rollup.config.js
@@ -5,7 +5,6 @@ import typescript from '@rollup/plugin-typescript'
 import jsx from 'acorn-jsx'
 import esbuild from 'rollup-plugin-esbuild'
 import postcss from 'rollup-plugin-postcss'
-import { sizeSnapshot } from 'rollup-plugin-size-snapshot'
 const createBabelConfig = require('./babel.config')
 
 const extensions = ['.js', '.ts', '.tsx']
@@ -54,13 +53,15 @@ function createDeclarationConfig(input, output) {
 function createESMConfig(input, output) {
   return {
     input,
-    output: { file: output, format: 'esm' },
+    output: [
+      { file: `${output}.js`, format: 'esm' },
+      { file: `${output}.mjs`, format: 'esm' },
+    ],
     external,
     plugins: [
       nodeResolve({ extensions }),
       postcss({ extract: 'style.css' }),
       getEsbuild('node12'),
-      sizeSnapshot(),
     ],
   }
 }
@@ -76,7 +77,6 @@ function createCommonJSConfig(input, output) {
       postcss({ extract: 'style.css' }),
       typescript(),
       babel(getBabelOptions({ ie: 11 })),
-      sizeSnapshot(),
     ],
   }
 }
@@ -85,7 +85,6 @@ export default function (args) {
   return [
     createDeclarationConfig('src/index.ts', 'dist'),
     createCommonJSConfig('src/index.ts', 'dist/index.js'),
-    createESMConfig('src/index.ts', 'dist/esm/index.mjs'),
-    createESMConfig('src/index.ts', 'dist/esm/index.js'),
+    createESMConfig('src/index.ts', 'dist/esm/index'),
   ]
 }